requirements.txt
streamlit==1.41.1
requests==2.32.3
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
